video accommodation
got video working for thumbnail clicks
with a new mediaCheck function to check if the thumbnail represents a
photo or video

then, tried to make it work with prevNext function for arrow/keyboard
navigation. this broke things…

experimented with hiding and showing the image and youtube overlays
either in the media check function or in the updateImage function.
Neither is working well yet.

first image you click seems to display properly, but the next one
includes both overlays. the next one after that displays properly, and
so on.

on the video, the href is going to the image overlay instead of the
youtube one now, although it worked before I tried to include
navigation.